requirements.txt
Package                  Version
------------------------ -----------
cachetools               5.3.3
certifi                  2024.2.2
charset-normalizer       3.3.2
google-api-core          2.19.0
google-api-python-client 2.130.0
google-auth              2.29.0
google-auth-httplib2     0.2.0
google-auth-oauthlib     1.2.0
googleapis-common-protos 1.63.0
httplib2                 0.22.0
idna                     3.7
numpy                    1.24.4
oauthlib                 3.2.2
pandas                   2.0.3
pip                      23.0.1
proto-plus               1.23.0
protobuf                 4.25.3
pyasn1                   0.6.0
pyasn1_modules           0.4.0
pyparsing                3.1.2
python-dateutil          2.9.0.post0
pytz                     2024.1
PyYAML                   6.0.1
requests                 2.32.2
requests-oauthlib        2.0.0
rsa                      4.9
setuptools               56.0.0
six                      1.16.0
tzdata                   2024.1
uritemplate              4.1.1
urllib3                  2.2.1
